refactor(webapp): relocate template-view-model + feature-local UI out of ui/ subfolder
