+++ frontend/projects/webapp/src/app/feature/budget/budget-list/create-budget/budget-creation-dialog.ts
@@ -22,8 +22,8 @@ import { MatSnackBar } from '@angular/material/snack-bar';
 import { MAT_DATE_FORMATS } from '@angular/material/core';
 import { MAT_DATE_FNS_FORMATS } from '@angular/material-date-fns-adapter';
 import { startOfMonth, setMonth, setYear } from 'date-fns';
-import { TemplatesList } from './ui/templates-list';
-import { type TemplateViewModel } from './ui/template-view-model';
+import { TemplatesList } from './templates-list';
+import { type TemplateViewModel } from './template-view-model';
 import { TemplateDetailsDialog } from './template-details-dialog';
 import { TemplateStore } from './services/template-store';
 import { ApiErrorLocalizer } from '@core/api/api-error-localizer';
