+++ frontend/projects/webapp/src/app/feature/budget/budget-list/create-budget/template-view-model.ts
